Move: StateController 에서 Money 객체로 분할하여 관련 메서드 이동
