+++ src/main/java/lotto/controller/StateController.java
@@ -3,6 +3,7 @@ package lotto.controller;
 import camp.nextstep.edu.missionutils.Randoms;
 import lotto.model.Lotto;
 import lotto.model.Lottos;
+import lotto.model.Money;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -13,11 +14,9 @@ import java.util.stream.Stream;
 
 public class StateController {
 
-    private static final int LOTTO_PRICE = 1000;
 
+    private Money money;
     private Lottos lottos;
-    private int money;
-    private int lottoCount;
     private Lotto answerLotto;
     private int bonus;
 
@@ -30,11 +29,8 @@ public class StateController {
     }
 
     private void enterMoney() {
-        int money = Integer.parseInt(InputView.readMoney());
         try {
-            validateMoney(money);
-            this.money = money;
-            this.lottoCount = money / LOTTO_PRICE;
+            money = new Money(Integer.parseInt(InputView.readMoney().trim()));
         } catch (IllegalArgumentException exception) {
             System.out.println(exception.getMessage());
             enterMoney();
@@ -43,16 +39,16 @@ public class StateController {
 
     private void purchaseLotto() {
         lottos = new Lottos(IntStream
-                .range(0, lottoCount)
+                .range(0, money.lottoCount())
                 .mapToObj(i -> new Lotto(Randoms.pickUniqueNumbersInRange(1, 45, 6)))
                 .toList());
-        OutputView.printLottoCount(lottoCount);
+        OutputView.printLottoCount(money.lottoCount());
         lottos.getLotto().forEach(lotto -> OutputView.printLotto(lotto.getNumbers()));
     }
 
     private void enterAnswer() {
         List<Integer> answerLotto = Stream
-                .of(InputView.readAnswer().split(","))
+                .of(InputView.readAnswer().trim().split(","))
                 .map(Integer::parseInt)
                 .toList();
         try {
@@ -65,20 +61,12 @@ public class StateController {
 
     private void enterBonus() {
         try {
-            validateBonus(Integer.parseInt(InputView.readBonus()));
+            int bonus = Integer.parseInt(InputView.readBonus().trim());
+            validateBonus(bonus);
             this.bonus = bonus;
         } catch (IllegalArgumentException exception) {
             System.out.println(exception.getMessage());
             enterBonus();
-        }
-    }
-
-    private void validateMoney(int money) {
-        if (money == 0) {
-            throw new IllegalArgumentException("[ERROR] 구입 금액은 1,000 원 이상이여야 합니다.");
-        }
-        if ((money % LOTTO_PRICE) != 0) {
-            throw new IllegalArgumentException("[ERROR] 구입 금액은 1,000 원 단위로 입력해주셔야 합니다.");
         }
     }
