+++ src/main/java/lotto/model/Lotto.java
@@ -29,5 +29,4 @@ public class Lotto {
             throw new IllegalArgumentException("[ERROR] 중복된 로또 번호는 입력할 수 없습니다.");
         }
     }
-    // TODO: 추가 기능 구현
 }
